minor tweaks

+++ src/engine/loader/detail/TextureLoader.cpp
@@ -13,7 +13,7 @@
 ecs::entity engine::detail::TextureLoader::load(ecs::registry &reg, std::string_view path, LoadingFlags flags)
 {
     int width = 0, height = 0, numChannels = 0;
-    stbi_set_flip_vertically_on_load(flags & LoadingFlags::MODEL_FLIP_TEXTURES);
+    stbi_set_flip_vertically_on_load(static_cast<int>(flags & LoadingFlags::MODEL_FLIP_TEXTURES));
     float *buff = stbi_loadf(path.data(), &width, &height, &numChannels, 0);
     if(!buff)
     {
@@ -45,7 +45,7 @@ ecs::entity engine::detail::TextureLoader::load(ecs::registry &reg, std::string_
 ecs::entity engine::detail::TextureLoader::load(ecs::registry &reg, std::size_t size, void const *data, LoadingFlags flags)
 {
     int width = 0, height = 0, numChannels = 0;
-    stbi_set_flip_vertically_on_load(flags & LoadingFlags::MODEL_FLIP_TEXTURES);
+    stbi_set_flip_vertically_on_load(static_cast<int>(flags & LoadingFlags::MODEL_FLIP_TEXTURES));
     float *buff = stbi_loadf_from_memory(static_cast<unsigned char const *>(data), size, &width, &height, &numChannels, 0);
     if(!buff)
     {
